implementing meshbuilder

+++ include/Render/MeshData.h
@@ -1,8 +1,7 @@
 #pragma once
 #include <vector>
-#include <stdint.h>
 
 struct MeshData {
 	std::vector<float> vertices;
-	std::vector<uint8_t> indices;
+	std::vector<unsigned int> indices;
 };
